# Spring_Backend/spring-backend-docs/openapi.yaml
openapi: 3.0.3
info:
  title: FL Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for the currently implemented federated learning backend endpoints.
    This document reflects the real API surface in the repository, not a future design.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /api/v1/clients:
    get:
      tags: [Clients]
      summary: List all hospitals/clients
      operationId: getAllClients
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HospitalDto'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/clients/me:
    get:
      tags: [Clients]
      summary: Get the current hospital record for the authenticated hospital user
      operationId: getMyClient
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current hospital details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HospitalDto'
        '400':
          description: Missing or invalid hospital id in JWT
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Hospital not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HospitalDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        contactEmail:
          type: string
          format: email
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri